pyproject.toml
[project]
name = "calibrationtools"
version = "0.1.1"
description = "Calibration runner framework to fit models to data."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "cfa-mrp>=0.0.4",
    "jsonschema>=4.26.0",
    "numpy>=2.3.4",
    "pyarrow>=23.0.0",
    "rich>=14.3.3",
    "scipy>=1.16.3",
]

[project.optional-dependencies]
cloud = ["cfa-cloudops @ git+https://github.com/CDCgov/cfa-cloudops.git"]

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["pytest>=9.0.2", "ty>=0.0.14", "scipy-stubs>=1.16.3.0"]
cloudops = ["cfa-cloudops @ git+https://github.com/CDCgov/cfa-cloudops.git"]

[tool.uv.workspace]
members = ["packages/*"]

[tool.pydoclint]
style = 'google'
exclude = '\.git|tests|packages'
